+++ gulp/flags.js
@@ -22,7 +22,8 @@ module.exports = function(options) {
             APPDOCK: argv.showAppdock || true,
             API_KEYS: argv.showApiKeys || false,
             CONTACT_MANAGEMENT: argv.contactManagement || true,
-            PLATFORM_REPORTING: argv.platformReporting || true
+            PLATFORM_REPORTING: argv.platformReporting || true,
+            EMAIL_PERMS: argv.emailPerms || false
           }
         }
       }))