+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
-- Bumping Flow Designer
+* Bumping Flow Designer
+* Email perms now under feature flag EMAIL_PERMS
 
 ## [4.10.23]
 * CXV1-8879 - Fixed category creation button on transfer lists contact panel